Few compilation warning fixed.
git-svn-id: https://svn.apache.org/repos/asf/harmony/enhanced/drlvm/trunk@523191 13f79535-47bb-0310-9956-ffa450edef68

+++ vm/jitrino/src/jet/cg_instr.cpp
@@ -443,7 +443,7 @@ void CodeGen::gen_write_barrier(JavaByteCodes opcode, Field_Handle fieldHandle)
         baseObject = Val(jobj, NULL_REF);
         rlock(baseObject);
         void* fieldAddress = field_get_address(fieldHandle);
-        slotAddress = Opnd(jobj, (jlong)(int_ptr)fieldAddress);
+        slotAddress = Opnd(jobj, (int_ptr)fieldAddress);
         rlock(slotAddress);
     }
     else if (opcode == OPCODE_AASTORE) {
